Split `EvalConstantInst` out of `eval.cpp`.
