+++ toolchain/check/BUILD
@@ -22,6 +22,7 @@ cc_library(
         "decl_name_stack.cpp",
         "deduce.cpp",
         "eval.cpp",
+        "eval_inst.cpp",
         "facet_type.cpp",
         "function.cpp",
         "generic.cpp",
@@ -57,6 +58,7 @@ cc_library(
         "deduce.h",
         "diagnostic_helpers.h",
         "eval.h",
+        "eval_inst.h",
         "facet_type.h",
         "function.h",
         "generic.h",
